Address S3 auth and CORS review
